90_TeamProject/01_PaperReview_240115/team_1/code/requirements.txt
pdfplumber
langchain-text-splitters
tiktoken
langchain_experimental
langchain_openai
langchain_core
langchain-community
streamlit
python-dotenv
PyYAML
python-magic
werkzeug